REFACTOR: file rename to reflect generic symbol inside
