+++ src/interceptors/model-validation/AccountExistsInterceptor.ts
@@ -1,7 +1,7 @@
 import { AccountModel } from "@/models/AccountModel";
 import { Inject, Interceptor } from "@tsed/di";
 import { MongooseModel } from "@tsed/mongoose";
-import { BaseRefExistsInterceptor } from "./BaseRefExistsInterceptor";
+import { BaseRefExistsInterceptor } from "./BaseRefExistsInterceptor`2";
 import { ExistenceInterceptorOpts } from "./ExistenceInterceptorOpts";
 
 @Interceptor()
